Split normalize flow logic out of sync_flow.go
Refactoring to split up #754
